refactor out format time as helper
