+++ src/helpers/index.ts
@@ -1,4 +1,5 @@
 import { addLeadingZero } from "./addLeadingZero";
+import { formatTime } from "./formatTime";
 import { sin, tan, cos, acos, atan2 } from "./mathFuncs";
 
-export { addLeadingZero, sin, tan, cos, acos, atan2 };
+export { addLeadingZero, formatTime, sin, tan, cos, acos, atan2 };
